♻️ refactor(recipes): rename use-case files to PascalCase
Rename all <name>.usecase.ts use-case files (and their specs) to the
PascalCase <Name>UseCase.ts convention, updating imports and barrels.
