+++ packages/recipes/src/application/adapter/RecipesAdapter.ts
@@ -39,18 +39,18 @@ import { IRecipesDelayedJobs } from '../../domain/jobs/IRecipesDelayedJobs';
 import { DeployRecipesJobFactory } from '../../infra/jobs/DeployRecipesJobFactory';
 import { UpdateRecipesAndGenerateSummariesJobFactory } from '../../infra/jobs/UpdateRecipesAndGenerateSummariesJobFactory';
 import { RecipesServices } from '../services/RecipesServices';
-import { CaptureRecipeUsecase } from '../useCases/captureRecipe/captureRecipe.usecase';
-import { CaptureRecipeWithPackagesUsecase } from '../useCases/captureRecipeWithPackages/captureRecipeWithPackages.usecase';
-import { DeleteRecipeUsecase } from '../useCases/deleteRecipe/deleteRecipe.usecase';
-import { DeleteRecipesBatchUsecase } from '../useCases/deleteRecipesBatch/deleteRecipesBatch.usecase';
-import { FindRecipeBySlugUsecase } from '../useCases/findRecipeBySlug/findRecipeBySlug.usecase';
-import { GetRecipeByIdUsecase } from '../useCases/getRecipeById/getRecipeById.usecase';
-import { GetRecipeVersionUsecase } from '../useCases/getRecipeVersion/getRecipeVersion.usecase';
-import { ListRecipesBySpaceUsecase } from '../useCases/listRecipesBySpace/listRecipesBySpace.usecase';
-import { ListRecipeVersionsUsecase } from '../useCases/listRecipeVersions/listRecipeVersions.usecase';
-import { UpdateRecipeFromUIUsecase } from '../useCases/updateRecipeFromUI/updateRecipeFromUI.usecase';
-import { UpdateRecipesFromGitHubUsecase } from '../useCases/updateRecipesFromGitHub/updateRecipesFromGitHub.usecase';
-import { UpdateRecipesFromGitLabUsecase } from '../useCases/updateRecipesFromGitLab/updateRecipesFromGitLab.usecase';
+import { CaptureRecipeUsecase } from '../useCases/captureRecipe/CaptureRecipeUsecase';
+import { CaptureRecipeWithPackagesUsecase } from '../useCases/captureRecipeWithPackages/CaptureRecipeWithPackagesUsecase';
+import { DeleteRecipeUsecase } from '../useCases/deleteRecipe/DeleteRecipeUsecase';
+import { DeleteRecipesBatchUsecase } from '../useCases/deleteRecipesBatch/DeleteRecipesBatchUsecase';
+import { FindRecipeBySlugUsecase } from '../useCases/findRecipeBySlug/FindRecipeBySlugUsecase';
+import { GetRecipeByIdUsecase } from '../useCases/getRecipeById/GetRecipeByIdUsecase';
+import { GetRecipeVersionUsecase } from '../useCases/getRecipeVersion/GetRecipeVersionUsecase';
+import { ListRecipesBySpaceUsecase } from '../useCases/listRecipesBySpace/ListRecipesBySpaceUsecase';
+import { ListRecipeVersionsUsecase } from '../useCases/listRecipeVersions/ListRecipeVersionsUsecase';
+import { UpdateRecipeFromUIUsecase } from '../useCases/updateRecipeFromUI/UpdateRecipeFromUIUsecase';
+import { UpdateRecipesFromGitHubUsecase } from '../useCases/updateRecipesFromGitHub/UpdateRecipesFromGitHubUsecase';
+import { UpdateRecipesFromGitLabUsecase } from '../useCases/updateRecipesFromGitLab/UpdateRecipesFromGitLabUsecase';
 
 const origin = 'RecipesAdapter';
