+++ packages/recipes/src/application/useCases/index.ts
@@ -1,16 +1,16 @@
 // Re-export all use cases
-export { CaptureRecipeUsecase } from './captureRecipe/captureRecipe.usecase';
-export { UpdateRecipesFromGitHubUsecase } from './updateRecipesFromGitHub/updateRecipesFromGitHub.usecase';
-export { UpdateRecipesFromGitLabUsecase } from './updateRecipesFromGitLab/updateRecipesFromGitLab.usecase';
-export { UpdateRecipeFromUIUsecase } from './updateRecipeFromUI/updateRecipeFromUI.usecase';
-export { DeleteRecipeUsecase } from './deleteRecipe/deleteRecipe.usecase';
-export { GetRecipeByIdUsecase } from './getRecipeById/getRecipeById.usecase';
-export { FindRecipeBySlugUsecase } from './findRecipeBySlug/findRecipeBySlug.usecase';
-export { ListRecipesBySpaceUsecase } from './listRecipesBySpace/listRecipesBySpace.usecase';
-export { ListRecipeVersionsUsecase } from './listRecipeVersions/listRecipeVersions.usecase';
-export { GetRecipeVersionUsecase } from './getRecipeVersion/getRecipeVersion.usecase';
-export { DeleteRecipesBatchUsecase } from './deleteRecipesBatch/deleteRecipesBatch.usecase';
-export { BaseUpdateRecipesFromWebhookUsecase } from './updateRecipesFromWebhook/BaseUpdateRecipesFromWebhook.usecase';
+export { CaptureRecipeUsecase } from './captureRecipe/CaptureRecipeUsecase';
+export { UpdateRecipesFromGitHubUsecase } from './updateRecipesFromGitHub/UpdateRecipesFromGitHubUsecase';
+export { UpdateRecipesFromGitLabUsecase } from './updateRecipesFromGitLab/UpdateRecipesFromGitLabUsecase';
+export { UpdateRecipeFromUIUsecase } from './updateRecipeFromUI/UpdateRecipeFromUIUsecase';
+export { DeleteRecipeUsecase } from './deleteRecipe/DeleteRecipeUsecase';
+export { GetRecipeByIdUsecase } from './getRecipeById/GetRecipeByIdUsecase';
+export { FindRecipeBySlugUsecase } from './findRecipeBySlug/FindRecipeBySlugUsecase';
+export { ListRecipesBySpaceUsecase } from './listRecipesBySpace/ListRecipesBySpaceUsecase';
+export { ListRecipeVersionsUsecase } from './listRecipeVersions/ListRecipeVersionsUsecase';
+export { GetRecipeVersionUsecase } from './getRecipeVersion/GetRecipeVersionUsecase';
+export { DeleteRecipesBatchUsecase } from './deleteRecipesBatch/DeleteRecipesBatchUsecase';
+export { BaseUpdateRecipesFromWebhookUsecase } from './updateRecipesFromWebhook/BaseUpdateRecipesFromWebhookUsecase';
 
 // Re-export all types from shared for backward compatibility
 export * from '@packmind/node-utils';
